+++ src/app/views/messages/sent-messages.component.ts
@@ -2,11 +2,11 @@ import { Component, OnInit } from '@angular/core';
 import {MessageService} from './services/message.service';
 import {NgForm} from '@angular/forms';
 import {MessagePaginator} from "./mesage.object.mapper";
-
+import* as swal from 'sweetalert';
 @Component({
   selector: 'app-sent-messages',
   templateUrl: './sent-messages.component.html',
-  styles: []
+
 })
 export class SentMessagesComponent implements OnInit {
   contact_message: any;
@@ -23,7 +23,7 @@ export class SentMessagesComponent implements OnInit {
               }
   public onSendmessage(f: NgForm){
         this.messageservice.sendTocontact(f.value ).subscribe(
-          () => {this.updateMessagePaginator()} );
+          () => {this.updateMessagePaginator(); swal("Message Sent to contact!", "success");} );
         f.reset();
          }
   public getPaginatedMessage(request_url) {
@@ -31,8 +31,28 @@ export class SentMessagesComponent implements OnInit {
   }
 
   public onDelete(message_id: number){
-      this.messageservice.deleteContactMessage(message_id).subscribe(
-        ()=> { this.updateMessagePaginator()}
-      );
+
+
+
+    swal({
+      title: "Are you sure?",
+      text: "Once deleted, you will not be able to recover this message!",
+      icon: "warning",
+      dangerMode: true,
+    })
+      .then((willDelete) => {
+        if (willDelete) {
+
+          this.messageservice.deleteContactMessage(message_id).subscribe(
+            ()=> { this.updateMessagePaginator()} );
+          swal("Message has been deleted!", {
+            icon: "success",
+          });
+        } else {
+          swal("Your message is safe!");
+        }
+      });
+
+
   }
 }
